Add a AnalyzerDriver.Create method
Creating an analyzer driver is rather complicated today - you need to create the right one for the language and you need to know to attach that to the compilation etc. In this change I'm just adding a public method called Create and hiding the constructor.

Also hiding a method called GetAnalyzerDiagnostic which was just used by a test in the IDE layer. The remaining public methods on this type either are resonable or are needed for the implementation of the IDE driver. We have some ideas to write the IDE driver in a way that might not require these post preview.
 (changeset 1345464)