Implements a spec refinement to how autoprops behave in constructors.
For accesses to an autoprop in a corresponding [static|instance] constructor -

(a) It should bind to the property, but the property should be treated as a readable/writable even if it is a readonly autoprop.
(b) The definite assignment behavior should be as if directly accessing the backing field.
(c) It should code gen to the property accessor (if one exists) or a field access (if not).

The most observable outcome of all this is that it is possible to assign a readonly autoprop in the constructor and that
in a struct instance constructor you no longer need to initialize whole struct before using any of its autoprops.

The following code is now valid and does what you think it does:

public struct S
{
    public int X{get;}
    public int Y{get;}

    public S()
    {
        X = 42;
        Y = X;
    }

    public static void Main()
    {
        S s = new S();
        System.Console.WriteLine(s.Y);
    }
}

.

 (changeset 1346315)