Delete AnalysisContext.RegisterSessionAnalyzer
The RegisterSessionAnalyzer on an AnalysisContext is a way for one session analyzer to register another session analyzer. We haven't needed such functionality till and currently this doesn't function either because it doesn't call the Initialize methods of the session analyzers. Removing this method for preview. If we get requests for this scenario, we can add this API back.

With this removed, all the Register* methods on context all register actions. This was the only one that registered an analyzer. (changeset 1346448)